ar-pal: check pal device ids before routing
Previously, device check is did in AHAL, while even
with same audio device, pal device can be different
with audio device due to shared backend check.
So later when audio device is to be restored, this
can be blocked by AHAL due to same device check, so
stream can keep on a different PAL device which
doesn't map to corresponding audio device.
Move the check from AHAL to PAL as PAL knows actual
PAL device enabled for the streams.

Change-Id: Ib85c8c8b8bbbbfa5825cac31e57eab55c7d39ed5