pal: Update the SVA state after SSR up
If sound model is loaded during SSR down and up, that time card
will be in offline state. Once SVA is loaded and its in active state,
the actual SSR event will be posted active state which is unhandled. So
update SVA state to SSR and handle SSR event during SSR up call.

Change-Id: I7be776a9877a942b01e920dbb942c74dd67caf2b