pal: Configure DevicePP EC MFC with CodecConfig
In Voip UC there is a requirement to configure the sample rate
of EC MFC same as that of MFC Mic in.

However for BT devices the EP sampling is not same as Device Data
rate. Hence we need to use Codec Config due to SWB. EP runs at a
different rate than Codec Output data.

Change-Id: I6f3b854f0eb23364048a7af66995b3c836407fde