pal: configuration of mux/demux for regular usecase in non-virtual port setup
Speaker and handset share the same device subgraph instance. There is also a
mux/demux module which is used for placing the correct channel on to Left
and Right. This configuration is important for virtual port solution during
concurrency where UPD + other concurrencies(music playback/voicecalls) are
running for the proper placement of upd signal(only L has valid data) to be
rendered on the earpiece. Mux/Demux is a tagged module which always needs
to have a configuration. This change allows configuration of mux/demux for
regular usecase in non-virtual port setup.

Change-Id: I66c589592eaaba4c2b7a422a39d49568f2cc70dd