pal: remove isAbrEnabled reset after failures in A2dp start
Setting the isAbrEnabled to false in case of start failure will cause
existing active stream to not call stopAbr in case of device stop.
In case of device start failures, we do not call startAbr and we do not
call device stop as well. Also stopAbr has checks to not process stopAbr
if abrRefCnt is 0.

Change-Id: I91037e2200abbced5dd0093bddc0510d40d6ce90
Signed-off-by: Anand Mohan <quic_anamoh@quicinc.com>