pal: Deregister VA device before stopping the SVA session
While SVA session stop is in progress and in parallel
Rx device is getting registered as part of Rx stream start,
EC reference device graph might be still set to SVA graph despite the
graph being in stopped state. This might lead to next graph
start failure for SVA as ec ref state and device state are
out of sync.
Deregister the VA device before stopping the SVA session
to fix the issue and make sure the echo reference is reset properly.

Change-Id: I019904b92b49b04aa3c5784f7010f51bcf259951