pal: Change volumeRamp parameters during pause/resume.
When device with higher volume is switched to speaker on lower volume,
then initial playback after device switch on speaker, after quick resume
can result in pop due to ramp down from higher volume to lower volume.
To resolve this, set ramp to 0 and volume to 0 before pause,and restore
the ramp and volume after resume, so that instead of ramping down the data,
now ramp up will happen on data from volume 0 to desired volume.

Change-Id: I4cbd4d22ec3fd1009ce83b8b227364cdb8132e45