PAL: Handle EC ref for different Rx device
If same stream is making call to set EC ref for different Rx device,
it will set if for first device for next upcoming devices, it will just
increment the counter and don't stop the previous and start for this new one.

Now, it will handle this case, and in case of failures while setting or resetting
it will decrement and increment counter respectively.

Change-Id: I8b81fd668ef200202f10eb0bbf6d166c7ff50e98