pal: restructure the code for mixer event callback registration
Existing code was relying on isPauseRegistartionDone flag to register
both mixer event callback as well as pause mixer event. If pause mixer
event registartion failed, it set the flag to false and later when it
has to deregister mixer event callback it would not do it as the flag
is already set to false.

Introduced another flag isMixerEventCbRegd to separately register and
deregister mixer event callback.

Change-Id: I98ac8fceb0fe4afd8728df822db87da7a91b5584
Signed-off-by: Anand Mohan <quic_anamoh@quicinc.com>