pal: update pal device for device switch in A2dpSuspend/Resume
Pal devices were not updated for implicit switch during a2dpSuspend and
a2dpResume. This was leading to incorrect routing whenever there was
explicit routing from APM to switch to a device which was incorrectly
present in pal device.

This change ensures proper update of pal devices during a2dpSuspend and
a2dpResume.

Change-Id: I4c60582a90a3ed2fbfa436302f6158da95f1388e
Signed-off-by: Anand Mohan <quic_anamoh@quicinc.com>