pal: fix routing switched back to upd device
When UPD stream is concurrent with multiple
streams on speaker, if one of the streams on
speaker goes away, during restoring devie, UPD
device can be selected as the candidate which
is wrong.
Skip checking UPD stream when there could be
remaining active streams on speaker, UPD device
restoration should only happen when UPD is
the last shared backend stream.

Change-Id: Iede13514497e8986daef922883a9c4d5133d19dd