StreamACD: unlock mutex before set ACD engine detection data
To prevent a potential deadlock when ACD needs to be paused
during detection events, unlock mutex before calling the
function SetEngineDetectionData and lock mutex afterward.
This will allow the thread that handles stopping ACD engine
to complete first because it could be blocked indefinitely
by the thread that parses event and notifies client.

Also add a resume case in ACD active state to handle this
corner case when ACD needs to be paused and resumed during
detection events.

Change-Id: If51d5f9dc389a5300b915284c61f34a7ff75ded5