PAL: Check if stream had any device before connect
If device disconnect-connect event happens before load, then no device
will get disconnected, as no device was associated with this stream.
But during connect cycle, we were not checking if stream had any device or not,
we directly try to connect current device, and this leads to null pointer
reference issue, as sm_cfg_ was not initialized yet.

Now, checking of sm_cfg_ is initialized or not, as it will be initialized,
during load, and after this only device get associated with the stream for
first time.

Change-Id: Ie5bf6c0596aa1e00982e7947102c3e8740cf9155